Refactored RemoteUserOperations to be common-util
