+++ warehouse/query-core/src/test/resources/datawave/query/QueryLogicFactory.xml
@@ -278,7 +278,7 @@
     
     <bean id="easyRoleManager" class="datawave.webservice.query.logic.EasyRoleManager" />
     
-    <util:list id="RemoteUserServices" value-type="datawave.webservice.common.remote.RemoteUserOperations">
+    <util:list id="RemoteUserServices" value-type="datawave.security.authorization.RemoteUserOperations">
    </util:list>
 
 </beans>
