+++ web-services/security/src/test/java/datawave/security/authorization/remote/ConditionalRemoteUserOperationsTest.java
@@ -5,10 +5,9 @@ import datawave.security.authorization.AuthorizationException;
 import datawave.security.authorization.DatawavePrincipal;
 import datawave.security.authorization.DatawaveUser;
 import datawave.security.authorization.SubjectIssuerDNPair;
-import datawave.security.authorization.remote.ConditionalRemoteUserOperations;
 import datawave.user.AuthorizationsListBase;
 import datawave.user.DefaultAuthorizationsList;
-import datawave.webservice.common.remote.RemoteUserOperations;
+import datawave.security.authorization.RemoteUserOperations;
 import datawave.webservice.query.Query;
 import datawave.webservice.query.cachedresults.CacheableQueryRow;
 import datawave.webservice.query.result.EdgeQueryResponseBase;
